refactor: 커스텀훅 분리 및 동적 스타일 클래스 분리
