+++ src/shared/Tab/components/index.ts
@@ -1,3 +1,4 @@
 export { default as TabHeader } from './TabHeader';
 export { default as TabItem } from './TabItem';
 export { default as TabViewer } from './TabViewer';
+export { default as TabBar } from './TabBar';
