+++ src/shared/Tab/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useTab } from './useTab';
+export { default as useTabBarStyle } from './useTabBarStyle';
